+++ clients/admin-ui/cypress/e2e/config-wizard-plus.cy.ts
@@ -1,5 +1,6 @@
 import { stubPlus } from "cypress/support/stubs";
 
+import { ADD_SYSTEMS_ROUTE } from "~/constants";
 import { ClusterHealth } from "~/types/api";
 
 /**
@@ -8,7 +9,7 @@ import { ClusterHealth } from "~/types/api";
  */
 const goToDataFlowScanner = () => {
   // Go through the initial config wizard steps
-  cy.visit("/add-systems");
+  cy.visit(ADD_SYSTEMS_ROUTE);
 
   // Select Runtime scanner to move to scan step.
   cy.getByTestId("add-systems");
@@ -23,11 +24,11 @@ describe("Config wizard with plus settings", () => {
   beforeEach(() => {
     cy.login();
     cy.intercept("GET", "/api/v1/organization/*", {
-      fixture: "organization.json",
+      fixture: "organizations/default_organization.json",
     }).as("getOrganization");
 
     cy.intercept("PUT", "/api/v1/organization**", {
-      fixture: "organization.json",
+      fixture: "organizations/default_organization.json",
     }).as("updateOrganization");
   });
 
@@ -42,7 +43,7 @@ describe("Config wizard with plus settings", () => {
           cluster_error: null,
         },
       });
-      cy.visit("/add-systems");
+      cy.visit(ADD_SYSTEMS_ROUTE);
       cy.getByTestId("add-systems");
 
       cy.wait("@getPlusHealth");
@@ -61,7 +62,7 @@ describe("Config wizard with plus settings", () => {
           cluster_error: null,
         },
       });
-      cy.visit("/add-systems");
+      cy.visit(ADD_SYSTEMS_ROUTE);
       cy.getByTestId("add-systems");
 
       cy.wait("@getPlusHealth");
@@ -74,7 +75,7 @@ describe("Config wizard with plus settings", () => {
 
     it("Can show the scanner as enabled and healthy", () => {
       stubPlus(true);
-      cy.visit("/add-systems");
+      cy.visit(ADD_SYSTEMS_ROUTE);
       cy.getByTestId("add-systems");
 
       cy.wait("@getPlusHealth");
@@ -238,7 +239,7 @@ describe("Config wizard with plus settings", () => {
       cy.wait("@putScanResults");
       cy.getByTestId("scan-results");
       cy.getByTestId("register-btn").click();
-      cy.visit("/add-systems");
+      cy.visit(ADD_SYSTEMS_ROUTE);
       cy.getByTestId("add-systems");
     });