+++ clients/admin-ui/src/features/common/Layout.tsx
@@ -46,7 +46,7 @@ const Layout = ({
     isValidNotificationRoute;
 
   return (
-    <Flex data-testid={title} direction="column" height="100vh">
+    <Flex data-testid={title} direction="column">
       <Head>
         <title>Fides Admin UI - {title}</title>
         <meta name="description" content="Privacy Engineering Platform" />
@@ -56,7 +56,7 @@ const Layout = ({
       {/* TODO: remove this in a future release (see https://github.com/ethyca/fides/issues/2844) */}
       <NotificationBanner />
       <NavTopBar />
-      <Flex as="main" px={9} py={10} gap="40px" height="100%">
+      <Flex as="main" flexGrow={1} padding={10} gap={10}>
         <Box flex={0} flexShrink={0}>
           <NavSideBar />
         </Box>