+++ clients/admin-ui/src/features/common/nav/v2/NavTopBar.tsx
@@ -14,11 +14,12 @@ export const NavTopBar = () => {
   return (
     <Flex
       as="nav"
+      height={12}
+      paddingX={10}
       gap={1}
-      px={10}
-      py={2}
+      flexShrink={0}
+      alignItems="center"
       borderBottom="1px"
-      borderTop="1px"
       borderColor="gray.100"
     >
       {nav.groups.map((group) => {