+++ clients/admin-ui/src/types/api/models/EmailDocsSchema.ts
@@ -2,6 +2,8 @@
 /* tslint:disable */
 /* eslint-disable */
 
+import type { AdvancedSettings } from "./AdvancedSettings";
+
 /**
  * EmailDocsSchema Secrets Schema for API Docs
  */
@@ -9,4 +11,5 @@ export type EmailDocsSchema = {
   third_party_vendor_name: string;
   recipient_email_address: string;
   test_email_address?: string;
+  advanced_settings: AdvancedSettings;
 };