+++ src/fides/api/ctl/routes/validate.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel
 from fides.api.ctl.routes.util import API_PREFIX
 from fides.api.ctl.utils.api_router import APIRouter
 from fides.api.ops.api.v1 import scope_registry
-from fides.api.ops.util.oauth_util import verify_oauth_client_cli
+from fides.api.ops.util.oauth_util import verify_oauth_client_prod
 from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,
@@ -63,7 +63,7 @@ router = APIRouter(tags=["Validate"], prefix=f"{API_PREFIX}/validate")
 @router.post(
     "/",
     dependencies=[
-        Security(verify_oauth_client_cli, scopes=[scope_registry.VALIDATE_EXEC])
+        Security(verify_oauth_client_prod, scopes=[scope_registry.VALIDATE_EXEC])
     ],
     status_code=status.HTTP_200_OK,
     response_model=ValidateResponse,