+++ src/fides/api/main.py
@@ -56,8 +56,6 @@ from fides.api.ops.common_exceptions import (
 from fides.api.ops.models.application_config import ApplicationConfig
 from fides.api.ops.schemas.analytics import Event, ExtraData
 from fides.api.ops.service.connectors.saas.connector_registry_service import (
-    load_registry,
-    registry_file,
     update_saas_configs,
 )
 
@@ -69,7 +67,7 @@ from fides.api.ops.service.saas_request.override_implementations import *
 from fides.api.ops.tasks.scheduled.scheduler import scheduler
 from fides.api.ops.util.cache import get_cache
 from fides.api.ops.util.logger import _log_exception
-from fides.api.ops.util.oauth_util import get_root_client, verify_oauth_client_cli
+from fides.api.ops.util.oauth_util import get_root_client, verify_oauth_client_prod
 from fides.api.ops.util.system_manager_oauth_util import (
     get_system_fides_key,
     get_system_schema,
@@ -135,9 +133,8 @@ def create_fides_app(
     fastapi_app.include_router(api_router)
 
     if security_env == "dev":
-        # This removes auth requirements for CLI-related endpoints
-        # and is the default
-        fastapi_app.dependency_overrides[verify_oauth_client_cli] = get_root_client
+        # This removes auth requirements for specific endpoints
+        fastapi_app.dependency_overrides[verify_oauth_client_prod] = get_root_client
         fastapi_app.dependency_overrides[
             verify_oauth_client_for_system_from_request_body_cli
         ] = get_system_schema
@@ -273,9 +270,8 @@ async def setup_server() -> None:
 
     logger.info("Validating SaaS connector templates...")
     try:
-        registry = load_registry(registry_file)
         db = get_api_session()
-        update_saas_configs(registry, db)
+        update_saas_configs(db)
         logger.info("Finished loading saas templates")
     except Exception as e:
         logger.error(