+++ tests/ops/api/v1/endpoints/test_user_permission_endpoints.py
@@ -14,6 +14,7 @@ from fides.api.ops.api.v1.scope_registry import (
     PRIVACY_REQUEST_READ,
     SAAS_CONFIG_READ,
     SCOPE_REGISTRY,
+    USER_PERMISSION_ASSIGN_OWNERS,
     USER_PERMISSION_CREATE,
     USER_PERMISSION_READ,
     USER_PERMISSION_UPDATE,
@@ -23,7 +24,14 @@ from fides.core.config import CONFIG
 from fides.lib.models.client import ClientDetail
 from fides.lib.models.fides_user import FidesUser
 from fides.lib.models.fides_user_permissions import FidesUserPermissions
-from fides.lib.oauth.roles import OWNER, ROLES_TO_SCOPES_MAPPING, VIEWER
+from fides.lib.oauth.roles import (
+    APPROVER,
+    CONTRIBUTOR,
+    OWNER,
+    ROLES_TO_SCOPES_MAPPING,
+    VIEWER,
+    VIEWER_AND_APPROVER,
+)
 from tests.conftest import generate_auth_header_for_user, generate_role_header_for_user
 
 
@@ -43,37 +51,12 @@ class TestCreateUserPermissions:
         response = api_client.post(url, headers=auth_header, json={})
         assert HTTP_403_FORBIDDEN == response.status_code
 
-    def test_create_user_permissions_invalid_scope(
-        self,
-        db,
-        api_client,
-        generate_auth_header,
-        user,
-        url,
-    ) -> None:
-        auth_header = generate_auth_header([USER_PERMISSION_CREATE])
-        user = FidesUser.create(
-            db=db,
-            data={"username": "user_1", "password": "test_password"},
-        )
-
-        body = {"user_id": user.id, "scopes": ["not a real scope"]}
-
-        response = api_client.post(url, headers=auth_header, json=body)
-        assert HTTP_422_UNPROCESSABLE_ENTITY == response.status_code
-        assert response.json()["detail"][0]["loc"] == ["body", "scopes", 0]
-        assert (
-            "value is not a valid enumeration member"
-            in response.json()["detail"][0]["msg"]
-        )
-        user.delete(db)
-
     def test_create_user_permissions_invalid_user_id(
         self, db, api_client, generate_auth_header
     ) -> None:
         auth_header = generate_auth_header([USER_PERMISSION_CREATE])
         user_id = "bogus_user_id"
-        body = {"user_id": user_id, "scopes": [PRIVACY_REQUEST_READ]}
+        body = {"user_id": user_id, "roles": [VIEWER]}
         response = api_client.post(
             f"{V1_URL_PREFIX}/user/{user_id}/permission", headers=auth_header, json=body
         )
@@ -81,7 +64,7 @@ class TestCreateUserPermissions:
         assert HTTP_404_NOT_FOUND == response.status_code
         assert permissions is None
 
-    def test_create_user_permissions_add_scopes_no_client_to_update(
+    def test_create_user_permissions_no_client_to_update(
         self, db, api_client, generate_auth_header
     ) -> None:
         auth_header = generate_auth_header([USER_PERMISSION_CREATE])
@@ -90,7 +73,7 @@ class TestCreateUserPermissions:
             data={"username": "user_1", "password": "test_password"},
         )
 
-        body = {"user_id": user.id, "scopes": [PRIVACY_REQUEST_READ]}
+        body = {"user_id": user.id, "roles": [APPROVER]}
         response = api_client.post(
             f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
         )
@@ -98,13 +81,14 @@ class TestCreateUserPermissions:
         response_body = response.json()
         assert HTTP_201_CREATED == response.status_code
         assert response_body["id"] == permissions.id
-        assert permissions.scopes == [PRIVACY_REQUEST_READ]
+        assert permissions.roles == [APPROVER]
         assert not user.client
         user.delete(db)
 
-    def test_create_user_permissions_add_scopes(
+    def test_create_user_permissions_add_scopes_are_ignored(
         self, db, api_client, generate_auth_header
     ) -> None:
+        """Test scopes in the request body are ignored - the UI may still be passing in scopes"""
         auth_header = generate_auth_header([USER_PERMISSION_CREATE])
         user = FidesUser.create(
             db=db,
@@ -120,7 +104,11 @@ class TestCreateUserPermissions:
         db.add(client)
         db.commit()
 
-        body = {"user_id": user.id, "scopes": [PRIVACY_REQUEST_READ]}
+        body = {
+            "user_id": user.id,
+            "roles": [APPROVER],
+            "scopes": [PRIVACY_REQUEST_READ],
+        }
         response = api_client.post(
             f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
         )
@@ -128,10 +116,7 @@ class TestCreateUserPermissions:
         response_body = response.json()
         assert HTTP_201_CREATED == response.status_code
         assert response_body["id"] == permissions.id
-        assert permissions.scopes == [PRIVACY_REQUEST_READ]
-        assert (
-            user.client.roles == []
-        ), "Roles not specified in request so they were overridden"
+        assert permissions.roles == [APPROVER]
         user.delete(db)
 
     def test_create_user_permissions_add_bad_role(
@@ -153,6 +138,43 @@ class TestCreateUserPermissions:
             in response_body["detail"][0]["msg"]
         )
 
+    def test_create_user_permissions_roles_are_an_empty_list(
+        self, db, api_client, generate_auth_header
+    ) -> None:
+        """If roles are an empty list, user will have no roles"""
+        auth_header = generate_auth_header([USER_PERMISSION_CREATE])
+        user = FidesUser.create(
+            db=db,
+            data={"username": "user_1", "password": "test_password"},
+        )
+
+        body = {"user_id": user.id, "roles": []}
+        response = api_client.post(
+            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
+        )
+        permissions = FidesUserPermissions.get_by(db, field="user_id", value=user.id)
+        response_body = response.json()
+        assert HTTP_201_CREATED == response.status_code
+        assert response_body["roles"] == []
+        assert permissions.roles == []
+        user.delete(db)
+
+    def test_create_user_permissions_no_role_key(
+        self, db, api_client, generate_auth_header
+    ) -> None:
+        """Roles key explicitly required"""
+        auth_header = generate_auth_header([USER_PERMISSION_CREATE])
+        user = FidesUser.create(
+            db=db,
+            data={"username": "user_1", "password": "test_password"},
+        )
+
+        body = {"user_id": user.id}
+        response = api_client.post(
+            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
+        )
+        assert HTTP_422_UNPROCESSABLE_ENTITY == response.status_code
+
     def test_create_user_permissions_add_roles(
         self, db, api_client, generate_auth_header
     ) -> None:
@@ -164,7 +186,7 @@ class TestCreateUserPermissions:
         client = ClientDetail(
             hashed_secret="thisisatest",
             salt="thisisstillatest",
-            scopes=[PRIVACY_REQUEST_READ],
+            scopes=[],
             roles=[],
             user_id=user.id,
         )
@@ -179,42 +201,74 @@ class TestCreateUserPermissions:
         response_body = response.json()
         assert HTTP_201_CREATED == response.status_code
         assert response_body["id"] == permissions.id
-        assert permissions.scopes == []
         assert permissions.roles == [VIEWER]
-        assert client.scopes == [], "No scopes in body so they were overridden"
+
+        assert client.roles == [VIEWER]
+        assert client.scopes == [], "User create flow doesn't override client scopes"
         user.delete(db)
 
-    def test_create_roles_on_permission_object_and_client(
-        self, db, api_client, generate_auth_header
-    ) -> None:
-        auth_header = generate_auth_header([USER_PERMISSION_CREATE])
-        user = FidesUser.create(
+    @pytest.mark.parametrize(
+        "acting_user,added_role,expected_response",
+        [
+            ("owner_user", APPROVER, HTTP_201_CREATED),
+            ("owner_user", VIEWER_AND_APPROVER, HTTP_201_CREATED),
+            ("owner_user", VIEWER, HTTP_201_CREATED),
+            ("owner_user", CONTRIBUTOR, HTTP_201_CREATED),
+            ("owner_user", OWNER, HTTP_201_CREATED),
+            ("contributor_user", APPROVER, HTTP_201_CREATED),
+            ("contributor_user", VIEWER_AND_APPROVER, HTTP_201_CREATED),
+            ("contributor_user", VIEWER, HTTP_201_CREATED),
+            ("contributor_user", CONTRIBUTOR, HTTP_201_CREATED),
+            ("contributor_user", OWNER, HTTP_403_FORBIDDEN),
+            ("viewer_user", APPROVER, HTTP_403_FORBIDDEN),
+            ("viewer_user", VIEWER_AND_APPROVER, HTTP_403_FORBIDDEN),
+            ("viewer_user", VIEWER, HTTP_403_FORBIDDEN),
+            ("viewer_user", CONTRIBUTOR, HTTP_403_FORBIDDEN),
+            ("viewer_user", OWNER, HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", APPROVER, HTTP_403_FORBIDDEN),
+            (
+                "viewer_and_approver_user",
+                VIEWER_AND_APPROVER,
+                HTTP_403_FORBIDDEN,
+            ),
+            ("viewer_and_approver_user", VIEWER, HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", CONTRIBUTOR, HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", OWNER, HTTP_403_FORBIDDEN),
+            ("approver_user", APPROVER, HTTP_403_FORBIDDEN),
+            ("approver_user", VIEWER_AND_APPROVER, HTTP_403_FORBIDDEN),
+            ("approver_user", VIEWER, HTTP_403_FORBIDDEN),
+            ("approver_user", CONTRIBUTOR, HTTP_403_FORBIDDEN),
+            ("approver_user", OWNER, HTTP_403_FORBIDDEN),
+        ],
+    )
+    def test_create_user_roles_permission_matrix(
+        self,
+        db,
+        acting_user,
+        added_role,
+        expected_response,
+        request,
+        api_client,
+    ):
+        """Test asserting what roles can be granted to other users given your current role
+        when assigning initial user permissions"""
+        acting_user = request.getfixturevalue(acting_user)
+        updated_user = FidesUser.create(
             db=db,
-            data={"username": "user_1", "password": "test_password"},
+            data={"username": "new_user", "password": "test_password"},
         )
-        client = ClientDetail(
-            hashed_secret="thisisatest",
-            salt="thisisstillatest",
-            scopes=[],
-            roles=[VIEWER],
-            user_id=user.id,
-        )
-        db.add(client)
-        db.commit()
 
-        body = {"user_id": user.id, "roles": [OWNER]}
-        response = api_client.post(
-            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
+        auth_header = generate_role_header_for_user(
+            acting_user, roles=acting_user.permissions.roles
         )
-        permissions = FidesUserPermissions.get_by(db, field="user_id", value=user.id)
-        response_body = response.json()
-        assert HTTP_201_CREATED == response.status_code
-        assert response_body["id"] == permissions.id
-        assert permissions.scopes == []
-        assert permissions.roles == [OWNER]
-        db.refresh(client)
-        assert client.roles == [OWNER]
-        user.delete(db)
+        body = {"user_id": updated_user.id, "roles": [added_role]}
+        response = api_client.post(
+            f"{V1_URL_PREFIX}/user/{updated_user.id}/permission",
+            headers=auth_header,
+            json=body,
+        )
+        assert response.status_code == expected_response
+        updated_user.delete(db)
 
 
 class TestEditUserPermissions:
@@ -233,32 +287,6 @@ class TestEditUserPermissions:
         response = api_client.put(url, headers=auth_header, json={})
         assert HTTP_403_FORBIDDEN == response.status_code
 
-    def test_edit_user_permissions_invalid_scope(
-        self,
-        db,
-        api_client,
-        generate_auth_header,
-        url,
-    ) -> None:
-        auth_header = generate_auth_header([USER_PERMISSION_UPDATE])
-        user = FidesUser.create(
-            db=db,
-            data={"username": "user_1", "password": "test_password"},
-        )
-
-        body = {"scopes": ["not a real scope"]}
-
-        response = api_client.put(
-            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
-        )
-        assert HTTP_422_UNPROCESSABLE_ENTITY == response.status_code
-        assert response.json()["detail"][0]["loc"] == ["body", "scopes", 0]
-        assert (
-            "value is not a valid enumeration member"
-            in response.json()["detail"][0]["msg"]
-        )
-        user.delete(db)
-
     def test_edit_user_permissions_invalid_user_id(
         self, db, api_client, generate_auth_header
     ) -> None:
@@ -270,18 +298,19 @@ class TestEditUserPermissions:
         )
 
         permissions = FidesUserPermissions.create(
-            db=db, data={"user_id": user.id, "scopes": [PRIVACY_REQUEST_READ]}
+            db=db, data={"user_id": user.id, "roles": [APPROVER]}
         )
-        body = {"id": permissions.id, "scopes": [PRIVACY_REQUEST_READ]}
+        body = {"id": permissions.id, "roles": [APPROVER]}
         response = api_client.put(
             f"{V1_URL_PREFIX}/user/{invalid_user_id}/permission",
             headers=auth_header,
             json=body,
         )
+        assert HTTP_404_NOT_FOUND == response.status_code
+
         permissions = FidesUserPermissions.get_by(
             db, field="user_id", value=invalid_user_id
         )
-        assert HTTP_404_NOT_FOUND == response.status_code
         assert permissions is None
         user.delete(db)
 
@@ -298,7 +327,6 @@ class TestEditUserPermissions:
             db=db,
             data={
                 "user_id": user.id,
-                "scopes": [PRIVACY_REQUEST_READ],
                 "roles": [VIEWER],
             },
         )
@@ -308,37 +336,36 @@ class TestEditUserPermissions:
             db,
             CONFIG.security.oauth_client_id_length_bytes,
             CONFIG.security.oauth_client_secret_length_bytes,
-            scopes=[PRIVACY_REQUEST_READ],
             roles=[VIEWER],
             user_id=user.id,
         )
 
-        updated_scopes = [PRIVACY_REQUEST_READ, SAAS_CONFIG_READ]
-        body = {"scopes": updated_scopes}
+        updated_roles = [CONTRIBUTOR]
+        body = {"roles": updated_roles}
         response = api_client.put(
             f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
         )
         response_body = response.json()
         assert HTTP_200_OK == response.status_code
-        assert response_body["scopes"] == updated_scopes
-        assert (
-            response_body["roles"] == []
-        ), "Roles removed as they were not specified in the request"
+        assert response_body["roles"] == updated_roles
         assert response_body["id"] == permissions_id
 
         client: ClientDetail = ClientDetail.get_by(db, field="user_id", value=user.id)
-        assert client.scopes == updated_scopes
-        assert client.roles == []
+        assert client.roles == [CONTRIBUTOR]
 
         db.refresh(permissions)
-        assert permissions.scopes == updated_scopes
-        assert permissions.roles == []
+        assert permissions.roles == [CONTRIBUTOR]
         assert permissions.id == permissions_id
 
         user.delete(db)
 
-    def test_edit_user_scopes(self, db, api_client, generate_auth_header) -> None:
-        auth_header = generate_auth_header([USER_PERMISSION_UPDATE])
+    def test_edit_user_scopes_are_ignored(
+        self, db, api_client, generate_auth_header
+    ) -> None:
+        """Scopes in request are ignored as you can no longer edit user scopes"""
+        auth_header = generate_auth_header(
+            [USER_PERMISSION_UPDATE, USER_PERMISSION_ASSIGN_OWNERS]
+        )
         user = FidesUser.create(
             db=db,
             data={"username": "user_1", "password": "test_password"},
@@ -348,49 +375,45 @@ class TestEditUserPermissions:
             db=db,
             data={
                 "user_id": user.id,
-                "scopes": [PRIVACY_REQUEST_READ],
                 "roles": [VIEWER],
             },
         )
-        permissions_id = permissions.id
 
         ClientDetail.create_client_and_secret(
             db,
             CONFIG.security.oauth_client_id_length_bytes,
             CONFIG.security.oauth_client_secret_length_bytes,
-            scopes=[PRIVACY_REQUEST_READ],
             roles=[VIEWER],
             user_id=user.id,
         )
 
-        updated_scopes = [PRIVACY_REQUEST_READ, SAAS_CONFIG_READ]
-        # Note: It is odd that we have the permissions id in the request body.
-        # I've made it optional as the UI sends it.  Verify we ignore it.
-        body = {"id": "invalid_id", "scopes": updated_scopes}
+        body = {
+            "id": permissions.id,
+            "scopes": [PRIVACY_REQUEST_READ],
+            "roles": [OWNER],
+        }
         response = api_client.put(
             f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
         )
         response_body = response.json()
         assert HTTP_200_OK == response.status_code
-        assert response_body["scopes"] == updated_scopes
-        assert (
-            response_body["roles"] == []
-        ), "Roles removed as they were not specified in the request"
-        assert response_body["id"] == permissions_id
+        assert response_body["roles"] == [OWNER]
 
         client: ClientDetail = ClientDetail.get_by(db, field="user_id", value=user.id)
-        assert client.scopes == updated_scopes
-        assert client.roles == []
+        assert (
+            client.scopes == []
+        ), "Assert client scopes are not updated via the user permissions update flow"
+        assert client.roles == [OWNER]
 
         db.refresh(permissions)
-        assert permissions.scopes == updated_scopes
-        assert permissions.roles == []
-        assert permissions.id == permissions_id
+        assert permissions.roles == [OWNER]
 
         user.delete(db)
 
     def test_edit_user_roles(self, db, api_client, generate_auth_header) -> None:
-        auth_header = generate_auth_header([USER_PERMISSION_UPDATE])
+        auth_header = generate_auth_header(
+            [USER_PERMISSION_UPDATE, USER_PERMISSION_ASSIGN_OWNERS]
+        )
         user = FidesUser.create(
             db=db,
             data={"username": "user_1", "password": "test_password"},
@@ -400,7 +423,6 @@ class TestEditUserPermissions:
             db=db,
             data={
                 "user_id": user.id,
-                "scopes": [PRIVACY_REQUEST_READ],
                 "roles": [VIEWER],
             },
         )
@@ -409,7 +431,6 @@ class TestEditUserPermissions:
             db,
             CONFIG.security.oauth_client_id_length_bytes,
             CONFIG.security.oauth_client_secret_length_bytes,
-            scopes=[PRIVACY_REQUEST_READ],
             roles=[VIEWER],
             user_id=user.id,
         )
@@ -420,20 +441,153 @@ class TestEditUserPermissions:
         )
         response_body = response.json()
         assert HTTP_200_OK == response.status_code
-        assert (
-            response_body["scopes"] == []
-        ), "Scopes removed as they were not specified in the request"
         assert response_body["roles"] == [OWNER]
 
         client: ClientDetail = ClientDetail.get_by(db, field="user_id", value=user.id)
-        assert client.scopes == []
+        assert (
+            client.scopes == []
+        ), "Assert client scopes are not updated via the user permissions update flow"
         assert client.roles == [OWNER]
 
         db.refresh(permissions)
-        assert permissions.scopes == []
         assert permissions.roles == [OWNER]
 
         user.delete(db)
+
+    def test_edit_user_roles_remove_role(
+        self, db, api_client, generate_auth_header
+    ) -> None:
+        auth_header = generate_auth_header(
+            [USER_PERMISSION_UPDATE, USER_PERMISSION_ASSIGN_OWNERS]
+        )
+        user = FidesUser.create(
+            db=db,
+            data={"username": "user_1", "password": "test_password"},
+        )
+
+        permissions = FidesUserPermissions.create(
+            db=db,
+            data={
+                "user_id": user.id,
+                "roles": [OWNER],
+            },
+        )
+
+        ClientDetail.create_client_and_secret(
+            db,
+            CONFIG.security.oauth_client_id_length_bytes,
+            CONFIG.security.oauth_client_secret_length_bytes,
+            roles=[VIEWER],
+            user_id=user.id,
+        )
+
+        body = {"id": permissions.id, "roles": []}
+        response = api_client.put(
+            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
+        )
+        response_body = response.json()
+        assert HTTP_200_OK == response.status_code
+        assert response_body["roles"] == []
+
+        client: ClientDetail = ClientDetail.get_by(db, field="user_id", value=user.id)
+        assert (
+            client.scopes == []
+        ), "Assert client scopes are not updated via the user permissions update flow"
+        assert client.roles == []
+
+        db.refresh(permissions)
+        assert permissions.roles == []
+
+        user.delete(db)
+
+    def test_edit_user_roles_request_no_role_key(
+        self, db, api_client, generate_auth_header
+    ) -> None:
+        auth_header = generate_auth_header(
+            [USER_PERMISSION_UPDATE, USER_PERMISSION_ASSIGN_OWNERS]
+        )
+        user = FidesUser.create(
+            db=db,
+            data={"username": "user_1", "password": "test_password"},
+        )
+
+        permissions = FidesUserPermissions.create(
+            db=db,
+            data={
+                "user_id": user.id,
+                "roles": [OWNER],
+            },
+        )
+
+        body = {"id": permissions.id}
+        response = api_client.put(
+            f"{V1_URL_PREFIX}/user/{user.id}/permission", headers=auth_header, json=body
+        )
+        assert HTTP_422_UNPROCESSABLE_ENTITY == response.status_code
+
+        db.refresh(permissions)
+        assert permissions.roles == [OWNER]
+
+        user.delete(db)
+
+    @pytest.mark.parametrize(
+        "acting_user,updating_role,updated_user,expected_response",
+        [
+            ("owner_user", APPROVER, "user", HTTP_200_OK),
+            ("owner_user", VIEWER_AND_APPROVER, "user", HTTP_200_OK),
+            ("owner_user", VIEWER, "user", HTTP_200_OK),
+            ("owner_user", CONTRIBUTOR, "user", HTTP_200_OK),
+            ("owner_user", OWNER, "user", HTTP_200_OK),
+            ("contributor_user", APPROVER, "user", HTTP_200_OK),
+            ("contributor_user", VIEWER_AND_APPROVER, "user", HTTP_200_OK),
+            ("contributor_user", VIEWER, "user", HTTP_200_OK),
+            ("contributor_user", CONTRIBUTOR, "user", HTTP_200_OK),
+            ("contributor_user", OWNER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_user", APPROVER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_user", VIEWER_AND_APPROVER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_user", VIEWER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_user", CONTRIBUTOR, "user", HTTP_403_FORBIDDEN),
+            ("viewer_user", OWNER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", APPROVER, "user", HTTP_403_FORBIDDEN),
+            (
+                "viewer_and_approver_user",
+                VIEWER_AND_APPROVER,
+                "user",
+                HTTP_403_FORBIDDEN,
+            ),
+            ("viewer_and_approver_user", VIEWER, "user", HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", CONTRIBUTOR, "user", HTTP_403_FORBIDDEN),
+            ("viewer_and_approver_user", OWNER, "user", HTTP_403_FORBIDDEN),
+            ("approver_user", APPROVER, "user", HTTP_403_FORBIDDEN),
+            ("approver_user", VIEWER_AND_APPROVER, "user", HTTP_403_FORBIDDEN),
+            ("approver_user", VIEWER, "user", HTTP_403_FORBIDDEN),
+            ("approver_user", CONTRIBUTOR, "user", HTTP_403_FORBIDDEN),
+            ("approver_user", OWNER, "user", HTTP_403_FORBIDDEN),
+        ],
+    )
+    def test_edit_user_roles_permission_matrix(
+        self,
+        acting_user,
+        updating_role,
+        updated_user,
+        expected_response,
+        request,
+        api_client,
+    ):
+        """Test asserting what roles can be granted to other users given your current role"""
+        acting_user = request.getfixturevalue(acting_user)
+        updated_user = request.getfixturevalue(updated_user)
+
+        auth_header = generate_role_header_for_user(
+            acting_user, roles=acting_user.permissions.roles
+        )
+        body = {"id": updated_user.permissions.id, "roles": [updating_role]}
+        response = api_client.put(
+            f"{V1_URL_PREFIX}/user/{updated_user.id}/permission",
+            headers=auth_header,
+            json=body,
+        )
+        assert response.status_code == expected_response
 
 
 class TestGetUserPermissions:
@@ -454,7 +608,7 @@ class TestGetUserPermissions:
     @pytest.fixture(scope="function")
     def permissions(self, db, user) -> FidesUserPermissions:
         return FidesUserPermissions.create(
-            db=db, data={"user_id": user.id, "scopes": [PRIVACY_REQUEST_READ]}
+            db=db, data={"user_id": user.id, "roles": [APPROVER]}
         )
 
     def test_get_user_permissions_not_authenticated(self, api_client, user):
@@ -525,45 +679,14 @@ class TestGetUserPermissions:
         assert HTTP_200_OK == response.status_code
         assert response_body["id"] == permissions.id
         assert response_body["user_id"] == user.id
-        assert response_body["scopes"] == [PRIVACY_REQUEST_READ]
-        assert response_body["roles"] == []
-        assert response_body["total_scopes"] == [PRIVACY_REQUEST_READ]
-
-    def test_get_user_permissions_outdated_scope(
-        self, db, api_client, user, auth_user, permissions
-    ) -> None:
-        user.permissions.scopes = [PRIVACY_REQUEST_READ, "OLD_DEPRECATED_SCOPE"]
-        user.permissions.save(db)
-
-        scopes = [USER_PERMISSION_READ]
-        ClientDetail.create_client_and_secret(
-            db,
-            CONFIG.security.oauth_client_id_length_bytes,
-            CONFIG.security.oauth_client_secret_length_bytes,
-            scopes=scopes,
-            user_id=auth_user.id,
-        )
-        auth_header = generate_auth_header_for_user(auth_user, scopes)
-
-        response = api_client.get(
-            f"{V1_URL_PREFIX}/user/{user.id}/permission",
-            headers=auth_header,
-        )
-        response_body = response.json()
-        assert HTTP_200_OK == response.status_code
-        assert response_body["id"] == permissions.id
-        assert response_body["user_id"] == user.id
-        assert response_body["scopes"] == [
-            PRIVACY_REQUEST_READ
-        ]  # Deprecated scope ignored
-        assert response_body["roles"] == []
-        assert response_body["total_scopes"] == [PRIVACY_REQUEST_READ]
+        assert response_body["roles"] == [APPROVER]
+        assert response_body["total_scopes"] == ROLES_TO_SCOPES_MAPPING[APPROVER]
 
     def test_get_user_with_no_permissions_as_root(
         self, db, api_client, auth_user, root_auth_header
     ):
         FidesUserPermissions.create(
-            db=db, data={"user_id": auth_user.id, "scopes": None, "roles": None}
+            db=db, data={"user_id": auth_user.id, "roles": None}
         )
 
         response = api_client.get(
@@ -571,25 +694,27 @@ class TestGetUserPermissions:
             headers=root_auth_header,
         )
         resp = response.json()
-        assert resp["scopes"] == []
         assert resp["roles"] == []
         assert resp["total_scopes"] == []
         assert resp["user_id"] == auth_user.id
 
     def test_get_current_user_permissions(self, db, api_client, auth_user) -> None:
         # Note: Does not include USER_PERMISSION_READ.
-        scopes = [PRIVACY_REQUEST_READ, SAAS_CONFIG_READ]
         ClientDetail.create_client_and_secret(
             db,
             CONFIG.security.oauth_client_id_length_bytes,
             CONFIG.security.oauth_client_secret_length_bytes,
-            scopes=scopes,
+            roles=[VIEWER],
             user_id=auth_user.id,
         )
-        auth_header = generate_auth_header_for_user(auth_user, scopes)
-        permissions = FidesUserPermissions.create(
-            db=db, data={"user_id": auth_user.id, "scopes": scopes}
+        auth_header = generate_auth_header_for_user(
+            auth_user, ROLES_TO_SCOPES_MAPPING[VIEWER]
         )
+
+        permissions = FidesUserPermissions.create(
+            db=db, data={"user_id": auth_user.id, "roles": [VIEWER]}
+        )
+        assert USER_PERMISSION_READ not in ROLES_TO_SCOPES_MAPPING[VIEWER]
 
         response = api_client.get(
             f"{V1_URL_PREFIX}/user/{auth_user.id}/permission",
@@ -599,9 +724,8 @@ class TestGetUserPermissions:
         assert HTTP_200_OK == response.status_code
         assert response_body["id"] == permissions.id
         assert response_body["user_id"] == auth_user.id
-        assert response_body["scopes"] == scopes
-        assert response_body["total_scopes"] == scopes
-        assert response_body["roles"] == []
+        assert response_body["total_scopes"] == ROLES_TO_SCOPES_MAPPING[VIEWER]
+        assert response_body["roles"] == [VIEWER]
 
     def test_get_current_root_user_permissions(
         self, api_client, oauth_root_client, root_auth_header
@@ -614,8 +738,6 @@ class TestGetUserPermissions:
         assert HTTP_200_OK == response.status_code
         assert response_body["id"] == oauth_root_client.id
         assert response_body["user_id"] == oauth_root_client.id
-        assert len(response_body["scopes"]) == len(SCOPE_REGISTRY)
-        assert response_body["scopes"] == SCOPE_REGISTRY
         assert response_body["roles"] == [OWNER]
         assert response_body["total_scopes"] == sorted(SCOPE_REGISTRY)
 
@@ -629,8 +751,6 @@ class TestGetUserPermissions:
         assert HTTP_200_OK == response.status_code
         assert response_body["id"] == oauth_root_client.id
         assert response_body["user_id"] == oauth_root_client.id
-        assert len(response_body["scopes"]) == len(SCOPE_REGISTRY)
-        assert response_body["scopes"] == SCOPE_REGISTRY
         assert response_body["roles"] == [OWNER]
         assert response_body["total_scopes"] == sorted(SCOPE_REGISTRY)
 
@@ -672,7 +792,6 @@ class TestGetUserPermissions:
             headers=auth_header,
         )
         resp = response.json()
-        assert resp["scopes"] == []
         assert resp["roles"] == [VIEWER]
         assert resp["user_id"] == auth_user.id
         assert resp["total_scopes"] == sorted(ROLES_TO_SCOPES_MAPPING[VIEWER])
@@ -680,7 +799,6 @@ class TestGetUserPermissions:
     def test_get_other_user_roles_as_root(
         self, db, api_client, auth_user, root_auth_header
     ):
-
         FidesUserPermissions.create(
             db=db, data={"user_id": auth_user.id, "roles": [VIEWER]}
         )
@@ -690,7 +808,6 @@ class TestGetUserPermissions:
             headers=root_auth_header,
         )
         resp = response.json()
-        assert resp["scopes"] == []
         assert resp["roles"] == [VIEWER]
         assert resp["user_id"] == auth_user.id
         assert resp["total_scopes"] == sorted(ROLES_TO_SCOPES_MAPPING[VIEWER])
@@ -723,7 +840,6 @@ class TestGetUserPermissions:
         )
         assert response.status_code == 200
         resp = response.json()
-        assert resp["scopes"] == []
         assert resp["roles"] == [VIEWER]
         assert resp["user_id"] == auth_user.id
         assert resp["total_scopes"] == sorted(ROLES_TO_SCOPES_MAPPING[VIEWER])